+++ src/o3d3xx_node.cpp
@@ -40,7 +40,7 @@ public:
   O3D3xxNode()
     : timeout_millis_(500),
       publish_viz_images_(false),
-      publish_numpy_cloud_(false),
+      publish_xyzi_image_(false),
       spinner_(new ros::AsyncSpinner(1))
   {
     std::string camera_ip;
@@ -53,7 +53,7 @@ public:
     nh.param("password", password, o3d3xx::DEFAULT_PASSWORD);
     nh.param("timeout_millis", this->timeout_millis_, 500);
     nh.param("publish_viz_images", this->publish_viz_images_, false);
-    nh.param("publish_numpy_cloud", this->publish_numpy_cloud_, false);
+    nh.param("publish_xyzi_image", this->publish_xyzi_image_, false);
 
     this->frame_id_ = ros::this_node::getName() + "_link";
 
@@ -79,7 +79,7 @@ public:
     this->conf_pub_ = it.advertise("/confidence", 1);
     this->good_bad_pub_ = it.advertise("/good_bad_pixels", 1);
     this->hist_pub_ = it.advertise("/hist", 1);
-    this->numpy_cloud_pub_ = it.advertise("/numpy_cloud", 1);
+    this->xyzi_image_pub_ = it.advertise("/xyzi_image", 1);
 
     //----------------------
     // Advertised services
@@ -128,7 +128,7 @@ public:
     cv::Mat depth_img;
     cv::Mat depth_viz_img;
     cv::Mat hist_img;
-    cv::Mat numpy_cloud_img;
+    cv::Mat xyzi_img;
     double min, max;
 
     while (ros::ok())
@@ -171,9 +171,9 @@ public:
                            confidence_img).toImageMsg();
       this->conf_pub_.publish(confidence);
 
-      if (this->publish_numpy_cloud_)
+      if (this->publish_xyzi_image_)
       {
-        numpy_cloud_img.create(depth_img.rows, depth_img.cols, CV_32FC4);
+        xyzi_img.create(depth_img.rows, depth_img.cols, CV_32FC4);
 
         int num_points = cloud->height * cloud->width;
         int col = 0;
@@ -189,7 +189,7 @@ public:
             if (col == 0)
               {
                 row += 1;
-                row_ptr = numpy_cloud_img.ptr<float>(row);
+                row_ptr = xyzi_img.ptr<float>(row);
               }
 
             row_ptr[matrix_col] = pt.x;
@@ -198,11 +198,11 @@ public:
             row_ptr[matrix_col + 3] = pt.intensity;
           }
 
-        sensor_msgs::ImagePtr numpy_cloud =
+        sensor_msgs::ImagePtr xyzi_image_msg =
           cv_bridge::CvImage(head,
                              sensor_msgs::image_encodings::TYPE_32FC4,
-                             numpy_cloud_img).toImageMsg();
-        this->numpy_cloud_pub_.publish(numpy_cloud);
+                             xyzi_img).toImageMsg();
+        this->xyzi_image_pub_.publish(xyzi_image_msg);
       }
 
       if (this->publish_viz_images_)
@@ -378,7 +378,7 @@ public:
 private:
   int timeout_millis_;
   bool publish_viz_images_;
-  bool publish_numpy_cloud_;
+  bool publish_xyzi_image_;
   std::unique_ptr<ros::AsyncSpinner> spinner_;
   o3d3xx::Camera::Ptr cam_;
   o3d3xx::FrameGrabber::Ptr fg_;
@@ -392,7 +392,7 @@ private:
   image_transport::Publisher conf_pub_;
   image_transport::Publisher good_bad_pub_;
   image_transport::Publisher hist_pub_;
-  image_transport::Publisher numpy_cloud_pub_;
+  image_transport::Publisher xyzi_image_pub_;
 
   ros::ServiceServer version_srv_;
   ros::ServiceServer dump_srv_;